+++ android/app/src/main/java/net/pengcook/android/data/remote/api/FeedService.kt
@@ -1,7 +1,7 @@
 package net.pengcook.android.data.remote.api
 
 import net.pengcook.android.data.model.feed.item.FeedItemResponse
-import net.pengcook.android.data.model.feed.step.RecipeStepResponse
+import net.pengcook.android.data.model.step.RecipeStepResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Path
